Added repositories for choice history and choice win. Other repositories modified.

+++ RPSSL.GameService/RPSSL.GameService.Domain/Models/ChoiceWin.cs
@@ -4,7 +4,7 @@ namespace RPSSL.GameService.Domain.Models;
 
 public class ChoiceWin : BaseEntity
 {
-	public short ChoiceId { get; set; }
-	public short BeatsChoiceId { get; set; }
+	public int ChoiceId { get; set; }
+	public int BeatsChoiceId { get; set; }
 	public required string ActionName { get; set; }
 }
